Creating a recorder app that can record incoming frames and write them to a file for later analysis
reusing a bunch of pw_ingest source config for this
